+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * The argument `optim_args` actually works now! Thank you to Jakob Madsen (#1).
 * An argument `flat` has been added to `profileCI()` to deal with cases where one or both of the limits is (minus or plus) infinity because the profile log-likelihood fails to drop below the level that defines the confidence interval.
+* A `logLikFn` method has been added for objects inheriting from class `nls` and an example added to the documentation of `profileCI()`. 
 
 ## Bug fixes and minor improvements